Moved MongoDB-specific implementations of DataRepository
They're own package is a little bit more fitting
